requirements.txt
langchain>=0.1.0
langchain-community>=0.0.38
langchain-openai>=0.0.2
openai>=1.30.1
faiss-cpu>=1.7.4
pypdf>=4.2.0
streamlit>=1.35.0
tiktoken>=0.6.0
pytesseract>=0.3.10
Pillow>=10.3.0
python-docx>=1.1.2
python-dotenv>=1.0.1
numpy>=1.24.0
anthropic